+++ packages/salesforcedx-vscode-apex/test/vscode-integration/commands/forceApexExecute.test.ts
@@ -6,10 +6,11 @@
  */
 import { ExecuteService } from '@salesforce/apex-node';
 import { getRootWorkspacePath } from '@salesforce/salesforcedx-utils-vscode/out/src';
+import { ChannelService } from '@salesforce/salesforcedx-utils-vscode/out/src/commands';
 import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import { expect } from 'chai';
 import * as path from 'path';
-import { createSandbox, SinonStub } from 'sinon';
+import { createSandbox, SinonSpy, SinonStub } from 'sinon';
 import * as vscode from 'vscode';
 import { channelService } from '../../../src/channels';
 import {
@@ -159,11 +160,11 @@ describe('Force Apex Execute', () => {
 
   describe('Format Execute Anonymous Response', () => {
     let outputStub: SinonStub;
-    let showOutputChannel: SinonStub;
+    let showChannelOutputStub: SinonSpy;
 
     beforeEach(() => {
       outputStub = sb.stub(channelService, 'appendLine');
-      showOutputChannel = sb.stub(channelService, 'showChannelOutput');
+      showChannelOutputStub = sb.spy(ChannelService.prototype, 'showChannelOutput');
     });
 
     it('should format result correctly for a successful execution', async () => {
@@ -193,9 +194,8 @@ describe('Force Apex Execute', () => {
       );
 
       await executor.run({ type: 'CONTINUE', data: {} });
-
+      expect(showChannelOutputStub.notCalled).to.be.true;
       expect(outputStub.firstCall.args[0]).to.equal(expectedOutput);
-      expect(showOutputChannel.calledOnce).to.be.true;
     });
 
     it('should format result correctly for a compilation failure', async () => {
@@ -220,9 +220,8 @@ describe('Force Apex Execute', () => {
       );
 
       await executor.run({ type: 'CONTINUE', data: {} });
-
+      expect(showChannelOutputStub.notCalled).to.be.true;
       expect(outputStub.firstCall.args[0]).to.equal(expectedOutput);
-      expect(showOutputChannel.calledOnce).to.be.true;
     });
 
     it('should format result correctly for a runtime failure', async () => {
@@ -256,9 +255,8 @@ describe('Force Apex Execute', () => {
       );
 
       await executor.run({ type: 'CONTINUE', data: {} });
-
+      expect(showChannelOutputStub.notCalled).to.be.true;
       expect(outputStub.firstCall.args[0]).to.equal(expectedOutput);
-      expect(showOutputChannel.calledOnce).to.be.true;
     });
   });